DjangoWithAI/api_gateway/api_gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6